Added extra checking of collider movement to ensure accurate collisions
If an object adjusted it's position or collider state in general after it checked for it's own collisions the collider would no longer be accurate for when other objects check against it. This makes sure that all colliders in the compositor are up to date before the instance is done processing.